Fix for mentioning RTL names
The current @mention (for user names and channel names) uses the regex `\b` to detect the name boundaries, however that token doesn't support unicode characters. This solution replaces the word boundary with regex that match end of the word with any of the following white space, colon, comma and dot.

However, ending a name with a dot still has a problem as it doesn't get a mention link and this fix does not fix it. Part of the problem is that dots are allowed in the name validation regex. So this problem need to be solved later.

Another thing that this fix adds is a description to the UTF8 Name Validation settings, to inform chat admins about the characters that they need to avoid using in their regexes to not break other chat functionalities that depends on special characters to denote special meanings.